Conflicts:
	data_base/isf_data_base/db_initializers/load_simrun_general.py
	simrun/reduced_model/spiking_output.py

+++ data_base/analyze/spike_detection.py
@@ -1,8 +1,6 @@
 from functools import partial
 import pandas as pd
 import dask
-import numpy as np
-
 from single_cell_parser.analyze.membrane_potential_analysis import simple_spike_detection
 
 
@@ -14,12 +12,13 @@ def _helper(x, threshold=0):
     '''reads out a voltage trace, so it can be fed into simple_spike_detection()
     and converts the result back to pd.Series, so the result can be concatenated 
     to a dask dataframe'''
-    t = x.index.values
+    t = x.index.values.astype(float)
     values = x.values
-    spikes = simple_spike_detection(t,
-                                    values,
-                                    mode='regular',
-                                    threshold=threshold)
+    spikes = simple_spike_detection(
+        t,
+        values,
+        mode='regular',
+        threshold=threshold)
     #print(len(spikes))
     return pd.Series({lv: x for lv, x in enumerate(spikes)})
 
@@ -28,5 +27,8 @@ def spike_detection(ddf, scheduler=None, threshold=0):
     fun = partial(_helper, threshold=threshold)
     '''this method expects a dask dataframe and returns a pandas dataframe containing the spikes'''
     dummy = dask.compute(*list(
-        map(dask.delayed(lambda x: x.apply(fun, axis=1)), ddf.to_delayed())),scheduler=scheduler)
+        map(
+            dask.delayed(lambda x: x.apply(fun, axis=1)), 
+            ddf.to_delayed())),
+                         scheduler=scheduler)
     return pd.concat(dummy)
